refactor: extract custom widgets to gui/widgets/
