+++ app.py
@@ -227,436 +227,20 @@ def _redact_diagnostics_toml(text: str) -> str:
         return "# [config redaction failed]\n"
 
 
-# ==========================================================
-# CUSTOM WIDGETS
-# ==========================================================
+from gui.widgets import (
+    DroppableLineEdit,
+    DroppablePlainTextEdit,
+    SwitchButton,
+    AdvancedFlagRow,
+    Card,
+    FormSection,
+    ElidingLabel,
+    BasePage,
+    NavItem,
+    NavGroup,
+    StatusCard,
+)
 
-
-class DroppableLineEdit(QLineEdit):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setAcceptDrops(True)
-
-    def dragEnterEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-
-    def dragMoveEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-
-    def dropEvent(self, event):
-        paths = [
-            url.toLocalFile() for url in event.mimeData().urls() if url.isLocalFile()
-        ]
-        if paths:
-            self.setText(paths[0])
-        event.acceptProposedAction()
-
-
-class DroppablePlainTextEdit(QPlainTextEdit):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setAcceptDrops(True)
-
-    def dragEnterEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-
-    def dragMoveEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-
-    def dropEvent(self, event):
-        paths = [
-            url.toLocalFile() for url in event.mimeData().urls() if url.isLocalFile()
-        ]
-        if paths:
-            self.setPlainText("\n".join(paths))
-        event.acceptProposedAction()
-
-
-class SwitchButton(QWidget):
-    toggled = Signal(bool)
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._checked = False
-        self.setFixedSize(38, 22)
-        self.setCursor(Qt.CursorShape.PointingHandCursor)
-
-    def isChecked(self):
-        return self._checked
-
-    def setChecked(self, checked):
-        if self._checked != checked:
-            self._checked = checked
-            self.toggled.emit(self._checked)
-            self.update()
-
-    def mouseReleaseEvent(self, event):
-        self.setChecked(not self._checked)
-
-    def paintEvent(self, event):
-        painter = QPainter(self)
-        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
-        rect = self.rect().adjusted(0, 0, -1, -1)
-        app = QApplication.instance()
-        theme = app.property("theme") if app else "dark"
-        is_light = str(theme) == "light"
-        if self._checked:
-            border_color = QColor("#0F766E") if is_light else QColor("#4FB3A4")
-            bg_color = QColor("#CCFBF1") if is_light else QColor("#17332F")
-            circle_color = QColor("#0F766E") if is_light else QColor("#4FB3A4")
-        else:
-            border_color = QColor("#9CA3AF") if is_light else QColor("#3A4045")
-            bg_color = QColor("#F8F9FA") if is_light else QColor("#1D2226")
-            circle_color = QColor("#6B7280") if is_light else QColor("#5B6267")
-        painter.setPen(QPen(border_color, 1))
-        painter.setBrush(QBrush(bg_color))
-        painter.drawRoundedRect(rect, 11, 11)
-        if self._checked:
-            circle_x = self.width() - 2 - 16
-        else:
-            circle_x = 2
-        painter.setPen(Qt.PenStyle.NoPen)
-        painter.setBrush(QBrush(circle_color))
-        painter.drawEllipse(circle_x, 2, 16, 16)
-
-
-class AdvancedFlagRow(QWidget):
-    def __init__(self, def_, parent=None):
-        super().__init__(parent)
-        self.def_ = def_
-
-        self.enable = QCheckBox(f"--{def_.flag}")
-        self.enable.setObjectName("AdvancedFlagEnable")
-        self.enable.setChecked(False)
-        tooltip = self.def_.hint or self.def_.label
-        self.enable.setToolTip(tooltip)
-
-        self.value_widget = self._create_value_widget()
-        self.value_widget.setEnabled(False)
-
-        self.enable.toggled.connect(self.value_widget.setEnabled)
-
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(12)
-        layout.addWidget(self.enable, 0)
-        layout.addWidget(self.value_widget, 1)
-
-    def _create_value_widget(self):
-        kind = self.def_.kind
-
-        if kind == "bool":
-            w = QComboBox()
-            w.addItems(["true", "false"])
-            w.setCurrentText("true" if self.def_.default is not False else "false")
-            return w
-
-        if kind == "enum":
-            w = QComboBox()
-            w.addItems(list(self.def_.options))
-            if self.def_.default is not None:
-                w.setCurrentText(str(self.def_.default))
-            return w
-
-        if kind == "int":
-            w = QSpinBox()
-            w.setRange(
-                self.def_.min_val if self.def_.min_val is not None else 0,
-                self.def_.max_val if self.def_.max_val is not None else 999999,
-            )
-            if isinstance(self.def_.default, int):
-                w.setValue(self.def_.default)
-            if self.def_.hint:
-                w.setToolTip(self.def_.hint)
-            return w
-
-        if kind == "duration_minutes":
-            w = QSpinBox()
-            w.setRange(1, 1440)
-            w.setSuffix(" minutes")
-            if isinstance(self.def_.default, int):
-                w.setValue(self.def_.default)
-            else:
-                w.setValue(20)
-            if self.def_.hint:
-                w.setToolTip(self.def_.hint)
-            return w
-
-        if kind == "lines_repeat":
-            w = QPlainTextEdit()
-            w.setPlaceholderText(self.def_.placeholder)
-            w.setMaximumHeight(80)
-            if self.def_.hint:
-                w.setToolTip(self.def_.hint)
-            return w
-
-        # text, extensions, csv_repeat, date_range
-        w = QLineEdit()
-        if self.def_.secret_env:
-            w.setEchoMode(QLineEdit.EchoMode.Password)
-        w.setPlaceholderText(self.def_.placeholder)
-        if self.def_.hint:
-            w.setToolTip(self.def_.hint)
-        return w
-
-    def get_value(self):
-        kind = self.def_.kind
-
-        if kind == "bool":
-            return self.value_widget.currentText() == "true"
-
-        if kind == "enum":
-            return self.value_widget.currentText()
-
-        if kind in ("int", "duration_minutes"):
-            return self.value_widget.value()
-
-        if kind == "lines_repeat":
-            return self.value_widget.toPlainText()
-
-        return self.value_widget.text()
-
-    def set_value(self, value):
-        kind = self.def_.kind
-
-        if kind == "bool":
-            self.value_widget.setCurrentText("true" if bool(value) else "false")
-
-        elif kind == "enum":
-            self.value_widget.setCurrentText(str(value or ""))
-
-        elif kind in ("int", "duration_minutes"):
-            try:
-                self.value_widget.setValue(int(value))
-            except Exception:
-                pass
-
-        elif kind == "lines_repeat":
-            self.value_widget.setPlainText(str(value or ""))
-
-        else:
-            self.value_widget.setText(str(value or ""))
-
-    def state(self) -> dict:
-        return {
-            "enabled": self.enable.isChecked(),
-            "value": self.get_value(),
-        }
-
-    def set_state(self, state: dict):
-        if not isinstance(state, dict):
-            return
-        self.enable.blockSignals(True)
-        self.value_widget.blockSignals(True)
-        try:
-            is_enabled = bool(state.get("enabled", False))
-            self.enable.setChecked(is_enabled)
-            val = state.get("value", self.def_.default)
-            if val is not None:
-                self.set_value(val)
-            self.value_widget.setEnabled(is_enabled)
-        finally:
-            self.enable.blockSignals(False)
-            self.value_widget.blockSignals(False)
-
-
-class Card(QFrame):
-    def __init__(self, title, required=False, parent=None):
-        super().__init__(parent)
-        self.setObjectName("Card")
-        self.layout = QVBoxLayout(self)
-        self.layout.setContentsMargins(24, 24, 24, 24)
-        self.layout.setSpacing(18)
-        title_layout = QHBoxLayout()
-        title_label = QLabel(title)
-        title_label.setObjectName("CardTitle")
-        title_layout.addWidget(title_label)
-        if required:
-            req_label = QLabel("Required")
-            req_label.setObjectName("ReqBadge")
-            title_layout.addWidget(req_label)
-        title_layout.addStretch()
-        self.layout.addLayout(title_layout)
-        self.layout.addSpacing(16)
-
-
-class FormSection(QFormLayout):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setLabelAlignment(Qt.AlignmentFlag.AlignLeft)
-        self.setRowWrapPolicy(QFormLayout.RowWrapPolicy.WrapLongRows)
-        self.setVerticalSpacing(16)
-        self.setHorizontalSpacing(24)
-
-    def add_row(self, label_text, widget, hint=""):
-        lbl = QLabel(label_text)
-        lbl.setObjectName("FieldLabel")
-        if isinstance(widget, QPlainTextEdit):
-            widget.setSizePolicy(
-                QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Preferred
-            )
-        elif isinstance(widget, QWidget):
-            widget.setSizePolicy(QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Fixed)
-        if hint:
-            container = QVBoxLayout()
-            container.setContentsMargins(0, 0, 0, 0)
-            container.setSpacing(4)
-            container.addWidget(widget)
-            hint_lbl = QLabel(hint)
-            hint_lbl.setObjectName("Hint")
-            container.addWidget(hint_lbl)
-            self.addRow(lbl, container)
-        else:
-            self.addRow(lbl, widget)
-
-
-class ElidingLabel(QLabel):
-    def __init__(self, text="", elide=Qt.TextElideMode.ElideMiddle, parent=None):
-        super().__init__(parent)
-        self._elide = elide
-        self._full = text
-        self.setWordWrap(False)
-        self._refresh()
-
-    def setText(self, text):
-        self._full = text
-        self._refresh()
-
-    def text(self):
-        return self._full
-
-    def _refresh(self):
-        fm = self.fontMetrics()
-        w = self.contentsRect().width() or self.width()
-        super().setText(
-            fm.elidedText(self._full, self._elide, w) if w > 0 else self._full
-        )
-
-    def sizeHint(self):
-        fm = self.fontMetrics()
-        return QSize(fm.horizontalAdvance(self._full) + 2, fm.lineSpacing())
-
-    def minimumSizeHint(self):
-        fm = self.fontMetrics()
-        return QSize(fm.horizontalAdvance("…") + 4, fm.lineSpacing())
-
-    def resizeEvent(self, e):
-        super().resizeEvent(e)
-        self._refresh()
-
-    def showEvent(self, e):
-        super().showEvent(e)
-        self._refresh()
-
-    def changeEvent(self, e):
-        super().changeEvent(e)
-        if e.type() in (QEvent.Type.FontChange, QEvent.Type.StyleChange):
-            self._refresh()
-
-
-class BasePage(QWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.main_layout = QVBoxLayout(self)
-        self.main_layout.setContentsMargins(0, 0, 0, 0)
-        self.scroll = QScrollArea()
-        self.scroll.setWidgetResizable(True)
-        self.main_layout.addWidget(self.scroll)
-        self.scroll_content = QWidget()
-        self.scroll_layout = QHBoxLayout(self.scroll_content)
-        self.scroll_layout.setContentsMargins(0, 0, 0, 0)
-        self.scroll_layout.setSizeConstraint(QLayout.SizeConstraint.SetMinimumSize)
-        self.container = QWidget()
-        self.container.setMaximumWidth(1480)
-        self.container.setSizePolicy(
-            QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Preferred
-        )
-        self.layout = QVBoxLayout(self.container)
-        self.layout.setContentsMargins(24, 24, 24, 24)
-        self.layout.setSpacing(24)
-        self.scroll_layout.addStretch()
-        self.scroll_layout.addWidget(self.container, 1)
-        self.scroll_layout.addStretch()
-        self.scroll.setWidget(self.scroll_content)
-
-    def addWidget(self, widget):
-        self.layout.addWidget(widget)
-
-    def addStretch(self):
-        self.layout.addStretch()
-
-
-class NavItem(QPushButton):
-    def __init__(self, text, icon, parent=None):
-        super().__init__(text, parent)
-        self.setObjectName("NavBtn")
-        self.setCheckable(True)
-        self.setCursor(Qt.CursorShape.PointingHandCursor)
-        if icon is not None and not icon.isNull():
-            self.setIcon(icon)
-            self.setIconSize(QSize(16, 16))
-
-
-class NavGroup(QWidget):
-    def __init__(self, title, items, parent=None):
-        super().__init__(parent)
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 16)
-        layout.setSpacing(4)
-        if title:
-            lbl = QLabel(title)
-            lbl.setObjectName("NavTitle")
-            layout.addWidget(lbl)
-        for item in items:
-            layout.addWidget(item)
-
-
-class StatusCard(QFrame):
-    _DOT = {
-        "ok": "#22C55E",
-        "warn": "#F59E0B",
-        "err": "#EF4444",
-    }
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setObjectName("StatusCard")
-        lay = QVBoxLayout(self)
-        lay.setContentsMargins(16, 16, 16, 16)
-        lay.setSpacing(8)
-        self.dot_b, self.txt_b = self._row(lay, "Binary: …")
-        self.dot_s, self.txt_s = self._row(lay, "Server: Not Set")
-        self.set_binary("err", "Binary: Checking…")
-        self.set_server("err", "Server: Not Set")
-
-    def _row(self, lay, text):
-        dot = QLabel()
-        dot.setFixedSize(10, 10)
-        txt = QLabel(text)
-        txt.setObjectName("StatusText")
-        r = QHBoxLayout()
-        r.setSpacing(8)
-        r.addWidget(dot, 0, Qt.AlignmentFlag.AlignVCenter)
-        r.addWidget(txt, 1)
-        r.addStretch()
-        lay.addLayout(r)
-        return dot, txt
-
-    def _paint(self, dot, state):
-        color = self._DOT.get(state, self._DOT["err"])
-        dot.setStyleSheet(f"background:{color};border-radius:5px;")
-
-    def set_binary(self, state, text):
-        self._paint(self.dot_b, state)
-        self.txt_b.setText(text)
-
-    def set_server(self, state, text):
-        self._paint(self.dot_s, state)
-        self.txt_s.setText(text)
 
 
 # ==========================================================
