+++ gui/widgets/eliding_label.py
@@ -1,0 +1,46 @@
+from PySide6.QtCore import QEvent, QSize, Qt
+from PySide6.QtWidgets import QLabel
+
+
+class ElidingLabel(QLabel):
+    def __init__(self, text="", elide=Qt.TextElideMode.ElideMiddle, parent=None):
+        super().__init__(parent)
+        self._elide = elide
+        self._full = text
+        self.setWordWrap(False)
+        self._refresh()
+
+    def setText(self, text):
+        self._full = text
+        self._refresh()
+
+    def text(self):
+        return self._full
+
+    def _refresh(self):
+        fm = self.fontMetrics()
+        w = self.contentsRect().width() or self.width()
+        super().setText(
+            fm.elidedText(self._full, self._elide, w) if w > 0 else self._full
+        )
+
+    def sizeHint(self):
+        fm = self.fontMetrics()
+        return QSize(fm.horizontalAdvance(self._full) + 2, fm.lineSpacing())
+
+    def minimumSizeHint(self):
+        fm = self.fontMetrics()
+        return QSize(fm.horizontalAdvance("…") + 4, fm.lineSpacing())
+
+    def resizeEvent(self, e):
+        super().resizeEvent(e)
+        self._refresh()
+
+    def showEvent(self, e):
+        super().showEvent(e)
+        self._refresh()
+
+    def changeEvent(self, e):
+        super().changeEvent(e)
+        if e.type() in (QEvent.Type.FontChange, QEvent.Type.StyleChange):
+            self._refresh()
