+++ gui/widgets/switch_button.py
@@ -1,0 +1,51 @@
+from PySide6.QtCore import Qt, Signal
+from PySide6.QtGui import QBrush, QColor, QPainter, QPen
+from PySide6.QtWidgets import QApplication, QWidget
+
+
+class SwitchButton(QWidget):
+    toggled = Signal(bool)
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self._checked = False
+        self.setFixedSize(38, 22)
+        self.setCursor(Qt.CursorShape.PointingHandCursor)
+
+    def isChecked(self):
+        return self._checked
+
+    def setChecked(self, checked):
+        if self._checked != checked:
+            self._checked = checked
+            self.toggled.emit(self._checked)
+            self.update()
+
+    def mouseReleaseEvent(self, event):
+        self.setChecked(not self._checked)
+
+    def paintEvent(self, event):
+        painter = QPainter(self)
+        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
+        rect = self.rect().adjusted(0, 0, -1, -1)
+        app = QApplication.instance()
+        theme = app.property("theme") if app else "dark"
+        is_light = str(theme) == "light"
+        if self._checked:
+            border_color = QColor("#0F766E") if is_light else QColor("#4FB3A4")
+            bg_color = QColor("#CCFBF1") if is_light else QColor("#17332F")
+            circle_color = QColor("#0F766E") if is_light else QColor("#4FB3A4")
+        else:
+            border_color = QColor("#9CA3AF") if is_light else QColor("#3A4045")
+            bg_color = QColor("#F8F9FA") if is_light else QColor("#1D2226")
+            circle_color = QColor("#6B7280") if is_light else QColor("#5B6267")
+        painter.setPen(QPen(border_color, 1))
+        painter.setBrush(QBrush(bg_color))
+        painter.drawRoundedRect(rect, 11, 11)
+        if self._checked:
+            circle_x = self.width() - 2 - 16
+        else:
+            circle_x = 2
+        painter.setPen(Qt.PenStyle.NoPen)
+        painter.setBrush(QBrush(circle_color))
+        painter.drawEllipse(circle_x, 2, 16, 16)
